+++ wiring/inc/spark_wiring_wifi.h
@@ -231,8 +231,8 @@ public:
 
     IPAddress resolve(const char* name)
     {
-        HAL_IPAddress ip;
-        return (inet_gethostbyname(name, strlen(name), &ip, *this, NULL)<0) ?
+        HAL_IPAddress ip = {0};
+        return (inet_gethostbyname(name, strlen(name), &ip, *this, NULL) != 0) ?
                 IPAddress(uint32_t(0)) : IPAddress(ip);
     }